refactor(banner-edit-components-name): components renamed to be more consistent
